# PERF: Optimize Allocations