--- a/SdpSerializer/Specifications/SdpVersion.cs
+++ b/SdpSerializer/Specifications/SdpVersion.cs
@@ -37,12 +37,7 @@ namespace Javor.SdpSerializer.Specifications
         /// <inheritdoc />
         public override string Encode()
         {
-            StringBuilder sb = new StringBuilder();
-            
-            sb.Append(SdpType);
-            sb.Append(Value);
-
-            return sb.ToString();
+            return $"{SdpType}{Value}";
         }
     }
 }