Hi @Jacfal, I've tried to make some performance optimizations to this repo. Following are the benchmark results before and after the changes:

**Benchmarks before changes:**
               Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
--------------------- |---------:|----------:|----------:|-------:|----------:|
 SdpEncodingBenchmark | 1.777 us | 0.0354 us | 0.0690 us | 0.3548 |    2.2 KB |

**Benchmarks after changes:**
               Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
--------------------- |---------:|----------:|----------:|-------:|----------:|
 SdpEncodingBenchmark | 1.596 us | 0.0266 us | 0.0249 us | 0.2480 |   1.55 KB |

As you can see the allocations have gone down. I've also made sure that all the unit tests pass following these changes as well. Could you please confirm whether or not you think my changes are valid?

Thank you!